feat: front end is there
so i now have everything in place that i want. tailwind-style styling without
actually using tailwind. nice. just gotta add htmx and we're there baby. next up
is going to be adding postgres. or maybe refactoring the templating stuff to be
in its own module. we'll get there when we get there. but probably the
refactoring first because the project is starting to get a little unwieldy

Signed-off-by: Matt Provost <mattprovost6@gmail.com>