feat: properly support paths
okay so the last commit was a misnomer. now i am actually done with the router.
i forgot what bug this fixes but it does fix one of them. i think you could like
insert subpaths after a path match or something like that. whatever this is a
pretty good router now

Signed-off-by: Matt Provost <mattprovost6@gmail.com>