feat: create setup secret
wow this was a lot easier than i thought. it should also support running
multiple instances at the same time, so each instance should all spit out the
correct secret to their logs. still need to make it not spit that out when the
cluster has been initialized, but we'll get there when we get there

Signed-off-by: Matt Provost <mattprovost6@gmail.com>