feat: finish setup page
holy shit we're here. the setup page is officially done. we have sessions. we
have the secret. this is super cool. it seems to work pretty well. i think i
gated everything properly with middleware but idk honestly. we'll have to see i
think. but otherwise, next up is to set up the info for the initial admin user
and then we can continue onto the next kinda stuff.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>