feat: add tokens
so i decided to create an api token system. preflight will contact runway
through api tokens and that will be that. still need to implement environments,
but we'll get there when we get there. also need a middleware to block for api
tokens but we'll also get there when we get there

Signed-off-by: Matt Provost <mattprovost6@gmail.com>