feat: properly handle deployment errors
this is nice. still want to overhaul how i handle errors in this project with
some more structured like all that, but this is fine for now. i also want to
auto update the deployment page with some more info about the deployment and
error info and all that, but we'll get there

Signed-off-by: Matt Provost <mattprovost6@gmail.com>