feat: custom image support
i genuinely cannot believe this is completed. what an unbelievably outstanding
feature to have implemented. this is genuinely huge. completely opaque and feels
like magic. necessary for wingmen, which is the next major feature. when that
one is done. i can start actually implementing my homelab. this is so huge

Signed-off-by: Matt Provost <mattprovost6@gmail.com>