feat: wingmen fully implemented part 2?
okay i think we're finally there. just gotta make it not return a config point
if there's already one in the state and we're done. but for that, i need to
communicate a module name through everything.

separately, i need to make all of this passing around of like user, environment,
module, service, etc part of the context. that will be way better to work with.
i also need to clean up the data models that i am passing around. theyre really
messy right now but i can clean that up. code is easy to change.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>