chore: move everything (i think) to use native contexts
and so begins the effort to clean everything up. first on the docket is making
use of the built in context stuff. i want to completely clean up all of the
interfaces and data models so that there are clean hand off points in converting
from one format to another, but we'll get there when we get there.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>