chore: consolidate like structures
i was getting super disoriented by all of the different structures. this makes
it make a bit more sense. the api stuff is all in the deployment.go file, which
might get its name changed or be moved somewhere else. then the module.go is
where the evaluation type of stuff is and config.go is where the runtime type of
stuff is. segmented much much better so i can start dissecting everything and
rebuild it to be nicer to work with

Signed-off-by: Matt Provost <mattprovost6@gmail.com>