feat: support storing secret
we can now store secrets. they are deployed as k8s resources. will still need to
implement the function to reference the secret but we'll get there. will be used
primarily for the environment variables, container configs, and in wingmen. i
think the wingmen should directly try and read the resource because then we're
not sending secret info in transit and k8s can determine the rbac stuff.

i think i want to revamp the deployment page and add pagination to the dashboard
next up tho. that way it is just kinda more convenient to develop stuff. but in
theory, we should be pretty much right there to start implementing the
cloudflared service stuff.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>