feat: revamp deployment page
holy canola. we're here. i keep saying that, but this is honestly kinda the
point that i care about. the ui is at a point where its actually kinda not too
bad. like it doesnt look incredible but it like works and it actually works
kinda well. i can just sit on the deployment page and let the deployment run
through and do all that and whatever. still have a little cleanup type of stuff
going forward but we'll see.

i think i do want to rip out middlewares sometime soon but i just idk i dont
wanna have to go through and rip it out. it wont be too much work tbh but i'll
see. but what i do need to do is make it so that i can inherit state from the
previous successful deployment. that will be extraordinarily useful when going
from one deployment to another. also need to get it to where it will redeploy
secrets if it needs to. small things but big progress.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>