feat: inherit previous state but maybe a little smarter?
idk this should make it a little better i guess but still need to handle secrets
a bit better. i dont want to send them back down and i dont want to decode them,
so ill need a way to just say "use the previous value". probably a checkbox that
makes it just use the previous value on the backend. i.e. inherit the state
value, which would be the exact same.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>