feat: support wingman functions
cool, might need to clean up the interface of how functions are defined, but
this is super sick. wingmen can now define functions and implement their things
and whatever and then badaboom badabing, we now have super dynamic configuration
that is beautiful and incredible and all that stuff. super duper sick.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>