chore: more functionality for wingman functions
implementing actual cloudflared tokens right now. there were a few places where
it was a little lacking. primarily, i need to be able to create and read secrets
from the wingman. i also need to have the user and environment so that i can get
the right secret for that, so now i pass that info over the wire.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>