feat: better canceling support
this should provide better support for canceling an in-progress deployment by
actually canceling the context. it will properly cancel all of that junk and
whatever and all that. it also adds some changes around the ticking so that it
will also error out if the context is ended so that im not sitting here waiting
and doing whatever while the context was just canceled.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>